requirements.txt
argcomplete==2.0.0
argon2-cffi==21.1.0
asgiref==3.4.1
attrs==21.2.0
autopep8==1.6.0
backcall==0.2.0
backports.entry-points-selectable==1.1.1
backports.zoneinfo==0.2.1
bcrypt==3.2.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.3
dj-database-url==0.5.0
Django==3.2.5
django-adminlte3==0.1.6
django-allauth==0.46.0
django-bootstrap-modal-forms==2.2.0
django-cors-headers==3.5.0
django-crispy-forms==1.12.0
django-dotenv==1.4.2
django-filter==21.1
django-heroku==0.3.1
django-location-field==2.1.0
django-phonenumber-field==6.1.0
django-rest-framework==0.1.0
django-rest-passwordreset==1.2.0
django-shortcuts==1.6
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.17.2
entrypoints==0.3
filelock==3.3.2
fonttools==4.28.3
future==0.18.2
GDAL @ file:///C:/Users/Masterspace/Documents/projects/Eboda/GDAL-2.4.1-cp37-cp37m-win_amd64.whl
geographiclib==1.52
geopy==2.2.0
ghp-import==2.0.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==1.7.0
importlib-resources==5.7.1
inflection==0.5.1
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mergedeep==1.3.4
mistune==0.8.4
mkdocs==1.3.0
mysql-connector-python==8.0.28
mysqlclient==2.1.0
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.6
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.4.60
packaging==21.3
pandocfilters==1.5.0
paramiko==2.8.0
parso==0.8.3
phonenumbers==8.12.50
phonenumberslite==8.12.46
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.4.0
ply==3.11
prometheus-client==0.12.0
prompt-toolkit==3.0.23
protobuf==3.20.1
psycopg==3.0.14
psycopg2==2.9.2
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodomex==3.11.0
Pygments==2.10.0
PyJWT==2.4.0
PyNaCl==1.4.0
pyparsing==3.0.6
pyrsistent==0.18.0
pysmi==0.3.4
pysnmp==4.4.6
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.19.2
python3-openid==3.2.0
PythonTurtle==0.3.2
pytz==2021.1
pywin32==302
pywinpty==1.1.6
PyYAML==5.4.1
pyyaml_env_tag==0.1
pyzmq==22.3.0
qtconsole==5.2.1
QtPy==1.11.3
requests==2.27.1
requests-oauthlib==1.3.0
Send2Trash==1.8.0
setuptools-scm==6.3.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.1
stripe==2.63.0
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tomli==1.2.2
tornado==6.1
traitlets==5.1.1
typing-extensions==3.10.0.2
tzdata==2022.1
uritemplate==3.0.1
urllib3==1.26.9
virtualenv==20.10.0
watchdog==2.1.7
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.1.0
widgetsnbextension==3.5.2
zipp==3.7.0
